add authentication and authorization

+++ orders/views.py
@@ -4,20 +4,31 @@ from rest_framework.response import Response
 
 from rest_framework import status
 
+from rest_framework.permissions import IsAuthenticated
+
+from drf_spectacular.utils import extend_schema
+
 from .models import Order, OrderItem
 from .serializers import OrderSerializer, OrderItemSerializer
 from .services import delete_pending_order
+from .permissions import IsOrderItemOwner, IsOrderOwner
 
 
 class OrderListCreateView(APIView):
+    permission_classes = [IsAuthenticated]
+
     def get(self, request):
-        orders = Order.objects.all()
+        orders = Order.objects.filter(created_by=request.user)
         serializer = OrderSerializer(orders, many=True)
 
         return Response(
             serializer.data
         )
 
+    @extend_schema(
+        request=OrderSerializer,
+        responses=OrderSerializer
+    )
     def post(self, request):
         serializer = OrderSerializer(data=request.data)
 
@@ -35,8 +46,14 @@ class OrderListCreateView(APIView):
 
 
 class OrderDetailView(APIView):
+    permission_classes = [IsAuthenticated, IsOrderOwner]
+
     def get(self, request, pk):
-        order = get_object_or_404(Order, pk=pk)
+        order = get_object_or_404(Order, 
+                                  pk=pk)
+
+        self.check_object_permissions(request, order)
+        
         serializer = OrderSerializer(order)
 
         return Response(
@@ -44,7 +61,11 @@ class OrderDetailView(APIView):
         )
 
     def patch(self, request, pk):
-        order = get_object_or_404(Order, pk=pk) 
+        order = get_object_or_404(Order,
+                                  pk=p) 
+
+        self.check_object_permissions(request, order)
+        
         serializer = OrderSerializer(order,
                                      data=request.data,
                                      partial=True)
@@ -61,7 +82,10 @@ class OrderDetailView(APIView):
         ) 
 
     def delete(self, request, pk):
-        order = get_object_or_404(Order, pk=pk)
+        order = get_object_or_404(Order, 
+                                  pk=pk)
+
+        self.check_object_permissions(request, order)
 
         try:
             delete_pending_order(order)
@@ -78,8 +102,11 @@ class OrderDetailView(APIView):
     
 
 class OrderItemListView(APIView):
+    permission_classes = [IsAuthenticated]
+
     def get(self, request):
-        order_items = OrderItem.objects.all()
+        order_items = OrderItem.objects.filter(order__created_by=request.user)
+
         serializer = OrderItemSerializer(order_items,
                                          many=True)
 
@@ -89,8 +116,14 @@ class OrderItemListView(APIView):
 
 
 class OrderItemDetailView(APIView):
+    permission_classes = [IsAuthenticated, IsOrderItemOwner]
+
     def get(self, request, pk):
-        order_item = get_object_or_404(OrderItem, pk=pk)
+        order_item = get_object_or_404(OrderItem, 
+                                       pk=pk)
+
+        self.check_object_permissions(request, order_item)
+
         serializer = OrderItemSerializer(order_item) 
 
         return Response(
